Update index.html
Clicking on Site Home is leading back to DSM homepage instead of going to actual main page. Seems like a "/" is missing  before index. Not completely sure if that's how this works so just making a new branch and testing the fix. If it doesn't work I'll just revert it back.